requirements.txt
annotated-types==0.7.0
anyio==4.6.0
DateTime==5.5
fastapi==0.115.0
idna==3.10
jose==1.0.0
passlib==1.7.4
psycopg2-binary==2.9.9
pydantic==2.9.2
pydantic_core==2.23.4
pytz==2024.2
setuptools==75.1.0
sniffio==1.3.1
starlette==0.38.6
typing_extensions==4.12.2
zope.interface==7.0.3
